requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
crispy-bootstrap5==0.6
Django==4.0.5
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.7
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
